add exception

+++ src/main/kotlin/com/matb/ordering/api/controllers/VendorController.kt
@@ -1,8 +1,12 @@
-package com.matb.ordering.api.Controllers
+package com.matb.ordering.api.controllers
 
+import com.matb.ordering.api.exceptions.FoodNotFoundException
+import com.matb.ordering.api.exceptions.UserAlreadyTakenException
+import com.matb.ordering.api.exceptions.UserNotFoundException
 import com.matb.ordering.api.models.entities.Food
 import com.matb.ordering.api.models.entities.Report
 import com.matb.ordering.api.models.entities.Vendor
+import com.matb.ordering.api.models.repositories.CustomerRepository
 import com.matb.ordering.api.models.repositories.FoodRepository
 import com.matb.ordering.api.models.repositories.ReportRepository
 import com.matb.ordering.api.models.repositories.VendorRepository
@@ -12,14 +16,11 @@ import com.matb.ordering.api.models.responses.VendorResponse
 import com.matb.ordering.api.models.toFood
 import com.matb.ordering.api.models.toVendor
 import com.matb.ordering.api.models.toVendorResponse
-import org.apache.tomcat.jni.Local
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.web.bind.annotation.*
 import java.time.LocalDate
-import java.time.YearMonth
-import javax.transaction.Transactional
 
 @RestController
 @RequestMapping("/api/vendor")
@@ -27,13 +28,14 @@ class VendorController(
         private val encoder: BCryptPasswordEncoder,
         private val vendorRepository: VendorRepository,
         private val foodRepository: FoodRepository,
-        private val reportRepository: ReportRepository
+        private val reportRepository: ReportRepository,
+        private val customerRepository: CustomerRepository
 ){
 
     @PostMapping
     fun createVendor(@RequestBody vendor: VendorRequest) : ResponseEntity<Any>{
         if (vendorRepository.existsByUsername(vendor.username)){
-            return ResponseEntity("Username is already taken!", HttpStatus.CONFLICT)
+            throw UserAlreadyTakenException("vendor", vendor.username)
         }
         vendor.password = encoder.encode(vendor.password)
         return ResponseEntity(vendorRepository.save(vendor.toVendor()), HttpStatus.CREATED)
@@ -49,36 +51,42 @@ class VendorController(
 
     @GetMapping("/{username}")
     fun getVendorByUsername(@PathVariable(name = "username", required = true) username: String): ResponseEntity<VendorResponse>{
+        if (!vendorRepository.existsByUsername(username)) {
+            throw UserNotFoundException("vendor", username)
+        }
         return ResponseEntity(
                 vendorRepository.findByUsername(username).get().toVendorResponse(),
                 HttpStatus.OK)
     }
 
     @PutMapping
-    fun updateVendor(@RequestBody vendor: VendorRequest) : ResponseEntity<Any>{
+    fun updateVendor(@RequestBody vendor: VendorRequest) : ResponseEntity<VendorResponse>{
         if (!vendorRepository.existsByUsername(vendor.username)){
-            return ResponseEntity("User not found!", HttpStatus.NOT_FOUND)
+            throw UserNotFoundException("vendor", vendor.username)
         }
         var oldVendor = vendorRepository.findByUsername(vendor.username).get()
         if (vendor.password != "") {
             oldVendor.password = encoder.encode(vendor.password)
         }
         oldVendor.name = vendor.name
-        return ResponseEntity(vendorRepository.save(oldVendor), HttpStatus.OK)
+        return ResponseEntity(oldVendor.toVendorResponse(), HttpStatus.OK)
     }
 
     @DeleteMapping("/{username}")
     fun deleteVendor(@PathVariable(name = "username") username: String) : ResponseEntity<Void> {
+        if (!vendorRepository.existsByUsername(username)) {
+            throw UserNotFoundException("vendor", username)
+        }
         vendorRepository.deleteById(vendorRepository.findByUsername(username).get().id!!)
         return ResponseEntity(HttpStatus.OK)
     }
 
     // Food
     @PostMapping("/{username}")
-    fun createFood(@PathVariable(name = "username", required = true) username: String, @RequestBody foodRequest: FoodRequest): ResponseEntity<Any> {
+    fun createFood(@PathVariable(name = "username", required = true) username: String, @RequestBody foodRequest: FoodRequest): ResponseEntity<Food> {
         var vendor = vendorRepository.findByUsername(username)
         if (!vendor.isPresent)
-            return ResponseEntity("Vendor not found", HttpStatus.NOT_FOUND)
+            throw UserNotFoundException("vendor", username)
         return ResponseEntity(
                 foodRepository.save(
                         foodRequest.toFood(vendor.get())), HttpStatus.OK)
@@ -87,21 +95,25 @@ class VendorController(
     @PutMapping("/food/{id}")
     fun updateFood(
             @PathVariable (value = "id") foodId: Int,
-            @RequestBody foodRequest: FoodRequest): ResponseEntity<Any> {
+            @RequestBody foodRequest: FoodRequest): ResponseEntity<Food> {
         var food = foodRepository.findById(foodId)
         if (!food.isPresent) {
-            return ResponseEntity("Food not found", HttpStatus.NOT_FOUND)
+            throw FoodNotFoundException(foodId)
         }
         var foodUpdate = foodRequest.toFood(food.get().vendor!!)
-        foodUpdate.id = foodId
-        foodUpdate.state = foodRequest.state
-        return ResponseEntity(foodRepository.save(foodUpdate), HttpStatus.OK)
+        var newfood = food.get()
+        newfood.state = foodRequest.state
+        newfood.name = foodRequest.name
+        newfood.price = foodRequest.price
+        newfood.image = foodRequest.image
+        return ResponseEntity(newfood, HttpStatus.OK)
     }
-
-
 
     @DeleteMapping("/food/{id}")
     fun deleteFood(@PathVariable(value = "id") foodId: Int): ResponseEntity<Void> {
+        if (foodRepository.existsById(foodId)) {
+            throw FoodNotFoundException(foodId)
+        }
         foodRepository.deleteById(foodId)
         return ResponseEntity(HttpStatus.OK)
     }
@@ -111,6 +123,9 @@ class VendorController(
             @PathVariable(name = "username", required = true) username: String,
             @RequestBody listFood: List<FoodRequest>
     ): String {
+        if (!vendorRepository.existsByUsername(username)) {
+            throw UserNotFoundException("vendor", username)
+        }
         var vendor = vendorRepository.findByUsername(username).get()
         for (item in listFood) {
             ResponseEntity(foodRepository.save(item.toFood(vendor)), HttpStatus.OK)
@@ -122,6 +137,10 @@ class VendorController(
     fun getReport(@PathVariable(name = "username", required = true) username: String,
                   @PathVariable(name = "yearmonth", required = true) yearmonth: String
     ):ResponseEntity<List<Report>> {
+        if (!vendorRepository.existsByUsername(username)) {
+            throw UserNotFoundException("vendor", username)
+        }
+
         var startDate = LocalDate.parse(yearmonth+"-01")
         var endDate = startDate.plusDays(startDate.lengthOfMonth().toLong()-1)
         return ResponseEntity(reportRepository.findAllByVendorAndDateBetween(username,startDate,endDate), HttpStatus.OK)
